added-feedback-toast-and-alert

+++ frontend/src/components/sections/deleteImageText.tsx
@@ -1,5 +1,10 @@
+"use client";
+
 import { FaTrash } from 'react-icons/fa';
 import { deleteSectionImageText } from '@/sections/actions/section.action';
+import Swal from 'sweetalert2';
+import { useAppDispatch } from '@/store';
+import { resetToast, setToast } from '@/store/slices/toast.slice';
 
 type Props = {
 	portfolioId: string;
@@ -7,12 +12,32 @@ type Props = {
 };
 
 const DeleteImageText: React.FC<Props> = ({ portfolioId, sectionId }) => {
+	const dispatch = useAppDispatch();
+
 	const handleDeleteImageText = async () => {
-		try {
-			await deleteSectionImageText(portfolioId, sectionId)
-		} catch (error) {
-			console.error('Error deleting image-text:', error);
-		}
+		Swal.fire({
+			title: "Are you sure?",
+			text: "You won't be able to revert this!",
+			icon: "warning",
+			showCancelButton: true,
+			confirmButtonColor: "#3085d6",
+			cancelButtonColor: "#d33",
+			confirmButtonText: "Yes, delete it!"
+		}).then(async (result) => {
+			if (result.isConfirmed) {
+				const data = await deleteSectionImageText(portfolioId, sectionId)
+
+				if ("error" in data) {
+					dispatch(setToast({ message: data.error, type: "error" }));
+				}
+
+				if ("message" in data) {
+					dispatch(setToast({ message: data.message, type: "success" }));
+				}
+
+				setTimeout(() => dispatch(resetToast()), 3000);
+			}
+		});
 	};
 
 	return (
